move numeric traits
